SSP 修复更新脚本的一些问题

+++ trunk/user/shadowsocks/scripts/update_chnlist.sh
@@ -10,11 +10,10 @@ logger -st "SSP[$$]Update" "开始更新白名单..."
 rm -f /tmp/chnlist_domain.txt
 curl -k -s --connect-timeout 5 --retry 3 \
 ${CHNLIST_URL:-"https://raw.githubusercontent.com/felixonmars/dnsmasq-china-list/master/accelerated-domains.china.conf"} | \
-awk -F/ '{print $2}' | sort -u >> /tmp/chnlist_domain.txt
+awk -F/ '{print $2}' | sort -u > /tmp/chnlist_domain.txt
 
 [ ! -d /etc/storage/chnlist/ ] && mkdir /etc/storage/chnlist/
-mv -f /tmp/chnlist_domain.txt /etc/storage/chnlist/chnlist_domain.txt
-
+[ -s /tmp/chnlist_domain.txt ] && mv -f /tmp/chnlist_domain.txt /etc/storage/chnlist/chnlist_domain.txt && \
 mtd_storage.sh save >/dev/null 2>&1
 
 [ "$(nvram get ss_enable)" == "1" ] && /usr/bin/shadowsocks.sh rednsconf &>/dev/null
